Update NM_003060.3 to 4 & add RSG test_variant290
NM_003060.3's mappings where removed in the upgrade of the UTA to the
VVTA, due to the current exon position set being newer than the
mappings. Given the issues around such data we are unlikely to try and
recover this old data from the UTA, instead, if we need to, importing
from known good archived alignment data would be better.