Add RSG fix ref aln for test_variant314
With clean data NM_012309.4 maps to NC_000011.9:g.7033*5439*G>A as for
NR_110766, not  7033*6423* as before. There are a min of 3 exon
position sets for NM_012309.4 in the old data 23 for blat GRCh37 26 for
normal GRCh37(and possibly 38 alt), and 25 for GRCh38. There is no exon
set loaded with 23 exons, and we have no way of knowing whether the
GRCh38 alt == the GRCh37 set or just has the same count. This is
particularly an issue as the UTA made it's own exon aln. Even using our
own new alignment would only give plausible results, but not guarantee
identity to the originally published alignments, and can't help with
missing exon position sets.This shows an example of why we will have to
load from the original, rather than "recover" data if we want to fix
this sort of broken data, for seq without mappings in the current RefSeq
alignments.