Switch mapping for revoked transcript test_var112
NM_198455 is remarked as "record is based on a third party annotation
inferred from cow AJ416457.1" and has now been removed/revoked by the
NCBI, as such VVTA won't return it as a result from a genome query.
Instead we use the transcript as a query, but then check the genome
mapping as before.