Load real prot, add RSG & LRG for test_variant285
Load real protein id for NM_001349798.1, now that multiple transcripts
have the same CDS MD5 for different proteins. This gets mishandled by
the hgvs code returning the CDS MD5 not a real id when a single real id
is not found. To fix this the transcript->prot linkage has been filled
in on the database side with hand gathered data for the old sequences
that don't get filled in by normal updates. This patch fits the test to
this new data.