Use psudo-VCF delins style to fix hgvs ins issues
Since the hgvs has issues mapping from pure "ins" to non "ins" spans
try to use psudo-VCF delins for mapping on "ins" variants if normal
mapping fails. This is a more limited form of the later fallback on
total mapping failure, needed to prevent loss of existing mappings when
data, that more closely matches the genome, is added.