Add db permissions for travis, manually for now
tx_similarity_v is a legacy view with performance issues, that we do not
want to use if we do not have to, however it is currently in use. Add
the permissions to the travis script, since this would require manual
tweaking by users it will be fixed in the db later, but repeating the
grant should not cause issues.