Tweak vv db update RSG URLs 3/3 update GRCh38
The RefSeqGene to GRCh38 mappings have moved, and are also no longer
available in the same, filtered, file type. Adjust the URL, as well as
switching to fetching from a base url in a more flexible manner, to
handle differing names on data update. At the same time adjust the
script to abort if the data is not available, rather than failing
silently or crashing later depending on the web response.