Completely replace the exon checking func
This fixes at least two bugs, the more fundamental of which is that the
original function did not actually directly check the exon boundaries,
just ran a mapping and looked for changes. This should be quicker, fail
even when the normal mapping functions would give a pass, (which seems
to happen sometimes even with bad exon boundaries) and only fail for the
specific problems that we want to check for.