Add n VS c test set for expanded repeats
This test set is conceptually the same as the existing C set, but twice
as long and done with transcripts where there is a pair of n and c
versions with the same within transcript sequence, and the same from
transcript start coordinates, for the section in question. This allows
us to test the offset c based coordinates against the more raw 1 based
non offset n data. This test set also includes more n+offset type
coordinates as opposed to the mainly n-offset in the original set along
with more mutli-base repeats, which adds to the test coverage.