Clean up unintuitive intronic code and some names
Although the variant position is still messed with we no longer convert
the variant into a pseudo-g type and back when we get an intronic
coordinate. This patch also cleans up the use of offset, without
specifiers, in variable and function names, which is ambiguous. In the
process we end up fixing a n type to c type coordinate conversion bug
and preparing for addition of 3' UTR handling in the process.