Remove intronic_or_utr variable, fix repeat check
intronic_or_utr would sometimes store the UTR status, but this was
actually unused, instead it was mostly used to store the transcript when
an intron was detected, as part of the handling for the now fixed abuse
of the reference variable.
The regular expression fixed missing bad characters in a repeat if one
good repeat character was included.