Attempt to reduce SeqRepo calls when shifting
We attempt to push gap variants both left and right when validating,
this used to result in up to 50 separate SeqRepo seq fetch calls. This
commit replaces this with 1 50 base SeqRepo call. In the cases where we
trigger gap movement attempts but only attempt 1 move Micro-benchmarking
puts the 50bp fetch at about 4% slower than the 1bp fetch, but this
commit will at least nearly half the time spent in SeqRepo seq fetch
calls in every other case.  The net effect on our input tests is a
~22% decrease in time taken to run, so this is probably a net gain.