Convert hgvs transcript variation to hgvs object
This does have one affect on the output, namely that for intronic
variants that are mapped as '=' against the original genome we
previously took them as '=' unconditionally but now compare the output
'alt' sequence turning '=' into indels when the reference bases differ
between the target mapping and the original ref. The affected user input
tests (in test_inputs.py) are test_variant138 and test_variant169.