Add test for fixed crashing issue
During development (but not in the main dev or released branch) we, for
a while, had a bug where the start of a variant could end up after the
end, during processing, which was not caught in our current tests. This
would then cause a crash. The bug was fixed during later development,
so add a test to prevent re-regression for these variants.