Fix database handles to take a unix socket
For MySQL this is an extra optional setting, for PostgreSQL this is
in fact handled by an overloading of the 'host', which is how the
PostgreSQL driver handles things internally. This limits us to setting
the directory for the file to be used, not the full file name, for
Postgres, and also requires a patch to vv_hgvs to be used, as the path
otherwise confuses the db-url format that is used by vv_hgvs.

Tests are only fully triggered when a socket file is specified, for now,
the whole test set probably needs tweaking to allow us to use a non
standard file location at some point too, in order to allow more of
these tests to run multiprocess and reduce the risk of clobbering the
original, but this is sufficient for now.