Fetch the RS TX ID->LRG TX ID links into python
The database overheads, for searches even with indexing are high, so
instead fetch the 1.6K set of links for this into a python dictionary.
This should be fine with the current state but may need re-examining if
the number does grow in the future.  This does slightly increase run
time for simpler queries in the case that only 1 variant hits the RefSeq
to LRG mapping, but for BRCA1 time taken on my test machine goes from 44
seconds to 4(with ~0.05s being the RS->LRG mapping fetch).