Reduce re.search use in refseq_catch
This tested as a surprisingly key function for dwell-time in profiling,
specifically due to re.search. This requires a somewhat more complex
handling for nc_(NC_(NM_ type bad variants to keep our ability to handle
them, but this should not be significantly more compute time for the
normal case (being an extra conditional on '(' in the remainder).