Update VV to work with added vv_hgvs relative accs
Adding relative accessions to vv_hgvs for transcript variants, for use
when mapped from the genome, or provided deliberately (eg for intronic
variation states), was needed to avoid failures due to differing ideas
of exon locations for differing genomic mappings of the same transcript.

This however has some consequences to the behaviour of the vv_hgvs 'evm'
or 'easy variant mapper' that need relevant changes to VV to account
for, as well as allowing better handling of user input with specified
genomic accessions for intronic locations. This change is needed for VV
to fully work with the changes, but also requires the vv_hgvs updates or
else it will case crashes.